Move pipe mount tests to e2e tests directory
