+++ e2e/cli/mount/operation/test_rm.py
